split device logic
